requirements.txt
pymongo==4.6.2
python-dotenv==0.15.0
# bidict==0.22.1
# click==8.1.3
# dnspython==2.3.0
# eventlet==0.33.3
Flask==2.2.3
# Flask-SocketIO==5.3.3
# greenlet==2.0.2
# itsdangerous==2.1.2
Jinja2==3.1.2
# MarkupSafe==2.1.2
# python-engineio==4.4.0
# python-socketio==5.8.0
# six==1.16.0
Werkzeug==2.2.3
gunicorn==20.1.0